// types/queuetype.ts
export type queueType =
  | "RANKED_SOLO_5x5"
  | "RANKED_FLEX_SR"
  | "RANKED_FLEX_TT"
  | "RANKED_TFT"
  | "RANKED_TT"
  | "RANKED_FLEX_5x5"
  | "RANKED_FLEX_3x3"
  | "RANKED_TFT_TURBO"
  | "RANKED_TFT_STANDARD"
  | "RANKED_TFT_DRAFT"
  | "RANKED_TFT_CLASH"
  | "RANKED_TFT_RANKED"
  | "RANKED_TFT_TURBO_CLASH"
  | "RANKED_TFT_STANDARD_CLASH"
  | "RANKED_TFT_DRAFT_CLASH"
  | "RANKED_TFT_RANKED_CLASH"
  | "RANKED_TFT_TURBO_5x5"
  | "RANKED_TFT_STANDARD_5x5"
  | "RANKED_TFT_DRAFT_5x5"
  | "RANKED_TFT_RANKED_5x5"
  | "RANKED_TFT_TURBO_3x3"
  | "RANKED_TFT_STANDARD_3x3"
  | "RANKED_TFT_DRAFT_3x3"
  | "RANKED_TFT_RANKED_3x3"
  | "RANKED_TFT_TURBO_4x4"
  | "RANKED_TFT_STANDARD_4x4"
  | "RANKED_TFT_DRAFT_4x4"
  | "RANKED_TFT_RANKED_4x4"
  | "RANKED_TFT_TURBO_6x6"
  | "RANKED_TFT_STANDARD_6x6"
  | "RANKED_TFT_DRAFT_6x6"
  | "RANKED_TFT_RANKED_6x6"
  | "RANKED_TFT_TURBO_7x7"
  | "RANKED_TFT_STANDARD_7x7"
  | "RANKED_TFT_DRAFT_7x7"
  | "RANKED_TFT_RANKED_7x7"
  | "RANKED_TFT_TURBO_8x8"
  | "RANKED_TFT_STANDARD_8x8"
  | "RANKED_TFT_DRAFT_8x8"
  | "RANKED_TFT_RANKED_8x8";